ref(discover) Move discover endpoints into an 'app'
Move discover code into a django app like incidents is. This should help
keep the various bits of discover closer together as it is going to grow
a bit in the upcoming weeks.

Refs SEN-810
