+++ src/sentry/api/bases/organization.py
@@ -112,16 +112,6 @@ class OrganizationAuthProviderPermission(OrganizationPermission):
     }
 
 
-class OrganizationDiscoverSavedQueryPermission(OrganizationPermission):
-    # Relaxed permissions for saved queries in Discover
-    scope_map = {
-        'GET': ['org:read', 'org:write', 'org:admin'],
-        'POST': ['org:read', 'org:write', 'org:admin'],
-        'PUT': ['org:read', 'org:write', 'org:admin'],
-        'DELETE': ['org:read', 'org:write', 'org:admin'],
-    }
-
-
 class OrganizationUserReportsPermission(OrganizationPermission):
     scope_map = {
         'GET': ['project:read', 'project:write', 'project:admin'],
