+++ src/sentry/deletions/__init__.py
@@ -41,8 +41,6 @@ def load_defaults():
     default_manager.register(models.CommitAuthor, BulkModelDeletionTask)
     default_manager.register(models.CommitFileChange, BulkModelDeletionTask)
     default_manager.register(models.Dashboard, ModelDeletionTask)
-    default_manager.register(models.DiscoverSavedQuery, BulkModelDeletionTask)
-    default_manager.register(models.DiscoverSavedQueryProject, BulkModelDeletionTask)
     default_manager.register(models.EnvironmentProject, BulkModelDeletionTask)
     default_manager.register(models.Event, defaults.EventDeletionTask)
     default_manager.register(models.EventMapping, BulkModelDeletionTask)
