+++ src/sentry/discover/endpoints/discover_saved_query_detail.py
@@ -3,15 +3,15 @@ from __future__ import absolute_import
 from rest_framework.response import Response
 from sentry.api.serializers import serialize
 from sentry.api.exceptions import ResourceDoesNotExist
-from sentry.api.bases.organization import OrganizationDiscoverSavedQueryPermission
 from sentry.api.bases import OrganizationEndpoint
 from sentry.api.bases.discoversavedquery import DiscoverSavedQuerySerializer
 from sentry import features
-from sentry.models import DiscoverSavedQuery
+from sentry.discover.endpoints.bases import DiscoverSavedQueryPermission
+from sentry.discover.models import DiscoverSavedQuery
 
 
-class OrganizationDiscoverSavedQueryDetailEndpoint(OrganizationEndpoint):
-    permission_classes = (OrganizationDiscoverSavedQueryPermission, )
+class DiscoverSavedQueryDetailEndpoint(OrganizationEndpoint):
+    permission_classes = (DiscoverSavedQueryPermission, )
 
     def get(self, request, organization, query_id):
         """
